fix: refining docs; tweaking usage of useWallet hook
