+++ template_content/package.json.jinja
@@ -36,7 +36,7 @@
     "@playwright/test": "^1.35.0",
     "playwright": "^1.35.0",
     {% endif -%}
-    "vite": "4.4.9"
+    "vite": "^4.4.9"
   },
   "dependencies": {
     "@walletconnect/modal-sign-html": "^2.6.1",
@@ -44,7 +44,7 @@
     "@blockshake/defly-connect": "^1.1.6",
     "@daffiwallet/connect": "^1.0.3",
     "@perawallet/connect": "^1.3.1",
-    "@txnlab/use-wallet": "^2.1.1",
+    "@txnlab/use-wallet": "^2.4.0",
     "algosdk": "^2.5.0",
     {% if use_daisy_ui -%}
     "daisyui": "^3.1.0",
