+++ tests_generated/test_all_default_parameters_off/.copier-answers.yml
@@ -6,14 +6,12 @@ algod_server: http://localhost
 algod_token: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
 author_email: None
 author_name: None
-cloud_provider: none
 ide_jetbrains: false
 ide_vscode: false
 indexer_port: 8980
 indexer_server: http://localhost
 indexer_token: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
 project_name: test_all_default_parameters_off
-use_daisy_ui: false
 use_eslint_prettier: false
 use_github_actions: false
 use_jest: false
