+++ tests_generated/test_all_default_parameters_on_netlify/.copier-answers.yml
@@ -7,7 +7,6 @@ algod_token: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
 author_email: None
 author_name: None
 cloud_provider: netlify
-ide_jetbrains: false
 ide_vscode: true
 indexer_port: 8980
 indexer_server: http://localhost
