+++ tests_generated/test_all_default_parameters_on_netlify/.github/workflows/release.yaml
@@ -61,4 +61,5 @@ jobs:
         env:
           NETLIFY_SITE_ID: ${{ secrets.NETLIFY_SITE_ID }}
           NETLIFY_AUTH_TOKEN: ${{ secrets.NETLIFY_AUTH_TOKEN }}
+          # Set your netlify project env variables on your github repository (see README for more info)
