+++ tests_generated/test_all_default_parameters_on_vercel/.copier-answers.yml
@@ -7,7 +7,6 @@ algod_token: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
 author_email: None
 author_name: None
 cloud_provider: vercel
-ide_jetbrains: false
 ide_vscode: true
 indexer_port: 8980
 indexer_server: http://localhost
