+++ tests_generated/test_all_default_parameters_on_vercel/.github/workflows/release.yaml
@@ -62,5 +62,5 @@ jobs:
 
       - name: Deploy Project Artifacts to Vercel
         run: vercel deploy --prebuilt --prod --token=${{ secrets.VERCEL_TOKEN }}
-        # Set your vercel project env variables on your site instance on https://vercel.com/dashboard
+        # Set your vercel project env variables on your github repository (see README for more info)
